Add four diagnostic connection errors
- ConnectionFailedError
- ConnectionTimeoutError
- SSLVerificationError
- ProxyConnectionError

The first one is the most generic of the bunch. It's the error that
is raised when is no better (or at least defined) error to raise. This
sucks, but unfortunately requests and in particular urllib3 errors are
terrible. They lack so much context that it is challenging to present
specific, actionable errors. For example, there's no way to deduce that
a urllib3 NewConnectionError is caused by DNS unless you inspect the
error chain (for socket.gaierror) or parse the raw error string (which
is too hacky for my taste).

The other three are a lot better, although they are admittedly still
rather vague. The SSLVerificationError can't discern whether it was
caused by an expired or self-signed certificate for example. To work
around this, I've opted to include the original error as additional
context (with some limited rewriting if possible).

Anyway, the main benefit here is that despite the coarseness, we're able
to provide at least some actionable advice through the diagnostic hints.
In addition, these errors are way more readable than whatever requests
or urllib3 spits out (often being caught, stringified and presented in
single-line screaming red by pip).