Remove docs/tests for --no-python-version-warning
This flag is way more prevalent than initially thought. Since this flag
is already a no-op, it's better to simply remove the deprecation, tests,
and docs (i.e., hide it from CLI help) for the flag. This way, we can
avoid unnecessary churn, but avoid/discourage new occurances of the flag.

The deprecation has achieved the goal of communicating that this flag is
unnecessary today.